backend/jobs/tests.py
from rest_framework.test import APIClient
from django.urls import reverse
from django.contrib.auth import get_user_model
from rest_framework import status
from django.test import TestCase
from .models import JobSearch
from unittest.mock import patch
import time
from requests.models import Response

User = get_user_model()

class JobSearchTests(TestCase):
    def setUp(self):
        self.client = APIClient()
        self.user = User.objects.create_user(
            username='testuser', password='testpassword'
        )
        self.client.login(username='testuser', password='testpassword')
    